Changes in command TCs due to user_steps migration
