+++ tests/e2e_tests/component/iot_mqtt/commands/send/setup.py
@@ -1,4 +1,5 @@
 import time, re, datetime, requests
+from common.steps import service_created_precond,device_with_commands_created_precond,check_status_info
 from lettuce import step, world
 from iotqautils.cbUtils import CBUtils
 from iotqautils.gtwRest import Rest_Utils_SBC
@@ -6,7 +7,7 @@ from common.user_steps import UserSteps
 from common.gw_mqtt_commands import mqtt_command
 from threading import Thread
 from iotqautils.gtwMeasures import Gw_Measures_Utils
-from common.gw_configuration import CBROKER_URL,CBROKER_HEADER,GW_HOSTNAME,IOT_PORT,PATH_MQTT_COMMAND,IOT_SERVER_ROOT,DEF_ENTITY_TYPE,TIMEOUT_COMMAND
+from common.gw_configuration import CBROKER_URL,CBROKER_HEADER,GW_HOSTNAME,IOT_PORT,PATH_MQTT_COMMAND,IOT_SERVER_ROOT,DEF_ENTITY_TYPE
 
 cb = CBUtils(instance=GW_HOSTNAME,port=IOT_PORT,path_update=PATH_MQTT_COMMAND)
 api = Rest_Utils_SBC(server_root=IOT_SERVER_ROOT+'/iot')
@@ -20,35 +21,7 @@ def envia_comando(service, entityData):
     assert req.ok, 'ERROR: ' + req.text
     world.req_text=req.json()
 
-# Steps
-@step('a service with name "([^"]*)" and protocol "([^"]*)" created')
-def service_created_precond(step, service_name, protocol):
-    if protocol:
-        world.service_name = service_name
-        world.protocol = protocol
-        user_steps.service_precond(service_name, protocol)
-
-@step('a device with device id "([^"]*)", device name "([^"]*)", endpoint "([^"]*)", protocol "([^"]*)", command name "([^"]*)" and command value "([^"]*)" created')
-def device_with_endpoint_created_precond(step, device_id, device_name, endpoint, protocol, cmd_name, cmd_value):
-    replaces = {
-        "#": "|"
-    }
-    for kreplace in replaces:
-        cmd_value = cmd_value.replace(kreplace,replaces[kreplace])
-    command=[
-             {
-              "name": cmd_name,
-              "type": 'command',
-              "value": cmd_value
-              }
-             ]
-    if endpoint:
-        if not "http://" in endpoint:
-            endpoint = CBROKER_URL + endpoint       
-    if cmd_value:
-        user_steps.device_precond(device_id, endpoint, protocol, command, device_name)
-        world.device_name=device_name
-    
+# Steps    
 @step('I send a command to the IoTAgent with service "([^"]*)", device "([^"]*)", command "([^"]*)", cmd_type "([^"]*)" and value "([^"]*)"')
 def send_command(step, service, device_id, cmd_name, cmd_type, value):
     world.cmd_type=cmd_type
@@ -118,17 +91,6 @@ def wait_timeout_period(step, timeout, cmd_type):
     world.st=st
     world.ts=ts
     
-@step('the command of device "([^"]*)" with response "([^"]*)" and status "([^"]*)" is received by context broker')
-def check_status_info(step, asset_name, response, status):
-    world.response={}
-    if status=='OK':
-        check_command_cbroker(asset_name, status, response)
-    else:
-        if response:
-            world.response=status
-            check_command_cbroker(asset_name, status)
-            check_NOT_command_cbroker(asset_name, response, "Info")
-
 @step('And I request the command status')
 def requests_command_status(step):
     commands = api.get_listCommands(world.service_name, world.asset_name)
@@ -171,87 +133,3 @@ def check_command_list_results(step, command_list, status):
                 assert i['status'] == status, 'ERROR: El comando {} no tiene el estado {}'.format(cmd,status)
                 break
         assert command_exists, 'ERROR: El comando {} no se encuentra en la lista de comandos'.format(cmd)    
-
-def check_command_cbroker(asset_name, status, response={}, entity_type={}):
-    time.sleep(1)
-    timeinstant=1
-    if response:
-        replaces = {
-                "#": "|"
-        }
-        for kreplace in replaces:
-            response = response.replace(kreplace,replaces[kreplace])
-    req =  requests.get(CBROKER_URL+"/last")
-    cmd_name=str(world.cmd_name)+"_status"
-    print "Voy a comprobar el STATUS del Comando: " + str(cmd_name)
-    resp = req.json()
-    assert req.headers[CBROKER_HEADER] == world.service_name, 'ERROR de Cabecera: ' + world.service_name + ' esperada ' + str(req.headers[CBROKER_HEADER]) + ' recibida'
-    print 'Compruebo la cabecera {} con valor {} en last'.format(CBROKER_HEADER,req.headers[CBROKER_HEADER])
-    contextElement = resp['contextElements'][0]
-    assetElement = contextElement['id']
-    #print 'Dispositivo {}'.format(assetElement)
-    valueElement = contextElement['attributes'][0]['value']
-    #print '{}.{}'.format(model_name,asset_name)
-    nameElement = contextElement['attributes'][0]['name']
-    #print 'Metadatas {}'.format(metasElement)
-    assert assetElement == "{}".format(asset_name), 'ERROR: id: ' + str(asset_name) + " not found in: " + str(contextElement)
-    typeElement = contextElement['type']
-    if entity_type:
-        ent_type = entity_type
-    else:
-        ent_type = DEF_ENTITY_TYPE
-    assert typeElement == ent_type, 'ERROR: ' + ent_type + ' type expected, ' + typeElement + ' received'
-    print 'TYPE: ' + str(typeElement)
-    assert nameElement == cmd_name, 'ERROR: ' + cmd_name + ' name expected, ' + nameElement + ' received'
-    assert status in valueElement, 'ERROR: ' + status + ' value expected, ' + valueElement + ' received'
-    assert contextElement['attributes'][0]['metadatas'][0]['name'] == "TimeInstant", 'ERROR: ' + str(contextElement['attributes'][0]['metadatas'][0])
-    assert check_timestamp(contextElement['attributes'][0]['metadatas'][0]['value']), 'ERROR: metadata: ' + str(world.st) + " not found in: " + str(contextElement['attributes'][0]['metadatas'][0])
-    if response:
-        cmd_name=str(world.cmd_name)+"_info"
-        print "Voy a comprobar el INFO del Comando: " + str(cmd_name)
-        valueElement = contextElement['attributes'][1]['value']
-        nameElement = contextElement['attributes'][1]['name']
-        assert nameElement == cmd_name, 'ERROR: ' + cmd_name + ' name expected, ' + nameElement + ' received'
-        assert response in valueElement, 'ERROR: ' + response + ' value expected, ' + valueElement + ' received'
-        assert contextElement['attributes'][1]['metadatas'][0]['name'] == "TimeInstant", 'ERROR: ' + str(contextElement['attributes'][1]['metadatas'][0])
-        assert check_timestamp(contextElement['attributes'][1]['metadatas'][0]['value']), 'ERROR: metadata: ' + str(world.st) + " not found in: " + str(contextElement['attributes'][1]['metadatas'][0])
-        timeinstant+=1
-    nameTime = contextElement['attributes'][timeinstant]['name']
-    assert nameTime == "TimeInstant", 'ERROR: ' + "TimeInstant" + ' name expected, ' + nameTime + ' received'
-    assert check_timestamp(contextElement['attributes'][timeinstant]['value']), 'ERROR: timestamp: ' + str(world.st) + " not found in: " + str(contextElement['attributes'][timeinstant])
-
-def check_NOT_command_cbroker(asset_name, response, cmd_type):
-    time.sleep(1)
-    if cmd_type == "Status":
-        req =  requests.get(CBROKER_URL+"/lastStatus")
-    else:
-        req =  requests.get(CBROKER_URL+"/lastInfo")
-    resp = req.json()
-    assert req.headers[CBROKER_HEADER] == world.service_name, 'ERROR de Cabecera: ' + world.service_name + ' esperada ' + str(req.headers[CBROKER_HEADER]) + ' recibida'
-    print 'Compruebo la cabecera {} con valor {} en last{}'.format(CBROKER_HEADER,req.headers[CBROKER_HEADER],cmd_type)
-    contextElement = resp['contextElements'][0]
-    assetElement = contextElement['id']
-    assert assetElement != "{}".format(asset_name), 'ERROR: device: ' + str(asset_name) + " found in: " + str(contextElement)
-    print "Command is NOT received"
-    resp = world.req_text
-    if world.code:
-        assert resp['errorCode']['code'] == world.code, 'ERROR: code error expected ' + str(world.code) + " received " + str(resp['errorCode']['code'])
-        if world.response:
-            assert resp['errorCode']['reasonPhrase'] == str(world.response), 'ERROR: text error expected ' + str(world.response) + " received " + resp['errorCode']['reasonPhrase']
-        else:
-            assert resp['errorCode']['reasonPhrase'] == response, 'ERROR: text error expected ' + response + " received " + resp['errorCode']['reasonPhrase']
-
-def check_timestamp (timestamp):
-    st = datetime.datetime.utcfromtimestamp(world.ts).strftime('%Y-%m-%dT%H:%M:%S')
-    if st in timestamp:
-        return True
-    else:
-        st = datetime.datetime.utcfromtimestamp(world.ts+1).strftime('%Y-%m-%dT%H:%M:%S')
-        if st in timestamp:
-            return True
-        else:
-            st = datetime.datetime.utcfromtimestamp(world.ts-1).strftime('%Y-%m-%dT%H:%M:%S')
-            if st in timestamp:
-                return True
-            else:
-                return False
