+++ tests/e2e_tests/component/iot_mqtt/commands/send/terrain.py
@@ -1,8 +1,8 @@
 from lettuce import before, after, world
-from common.user_steps import UserSteps
+from common.functions import Functions
 
 
-user_steps = UserSteps()
+functions = Functions()
 
 
 @before.all
@@ -14,4 +14,4 @@ def setup():
 
 @after.all
 def clean(total):
-    user_steps.clean(world.remember)
+    functions.clean(world.remember)
